Add nbDaysSold per project and nbDaysAffected per week on person page. Add nbDaysAffected per person on projects list page. Remove useless ids in views. nbDaysSold is now a float.

+++ src/main/java/fr/xinta/atemia/web/EditProject.java
@@ -51,7 +51,7 @@ public class EditProject extends AbstractServlet {
             request.setAttribute("project", project);            
             
             try {
-                project.setNbDaysSold(Integer.parseInt(request.getParameter("nbDaysSold")));
+                project.setNbDaysSold(Float.parseFloat(request.getParameter("nbDaysSold")));
                 if (project.getNbDaysSold() < 0) {
                     throw new NumberFormatException();
                 }
